ftpkg-to-bmodel-generator/requirements.txt
Jinja2==3.1.4
PyYAML==6.0.2
jsonschema==4.22.0
py-solc-x==2.0.3
